put scoreboard logic into a separate file, left only render
